# lumina-api/swagger.yaml
openapi: 3.0.3
info:
  title: Lumina Bot API
  version: 3.0.0
  contact:
    name: Lumina Dev Team
    url: https://bot.luminasink.com
  description: |
    # Lumina Bot API

    REST API for the Lumina Bot dashboard, inventory/gacha system, Discord guild
    management, and the internal bot ↔ API bridge. Built on Node.js/Express,
    deployed to Vercel (production) and `localhost:3000` (dev).

    ---

    ## Authentication

    The API supports **two** authentication methods for user-facing routes:

    1. **Cookie `lumina_token` (preferred for browsers)**
       - `httpOnly`, `Secure` (prod), `SameSite=Lax`, max-age = 1 hour
       - Set by `POST /expapi/v1/login` and the OAuth2 callback
       - Sent automatically by the browser — **immune to XSS** (JavaScript cannot read it)

    2. **Header `Authorization: Bearer <jwt>` (legacy / non-browser)**
       - For CLI tools, scripts, integration tests, and any client that does not
         manage cookies
       - Same JWT, same 1h expiry

    JWT expiry: **1 hour**. There is no refresh-token rotation yet — clients must
    re-authenticate when the token expires.

    The session helper `verifyRequestAuth(req)` reads the cookie first, then
    falls back to the Bearer header. Either is accepted on every protected route.

    ---

    ## CSRF Protection

    State-changing routes (`POST`/`PUT`/`DELETE` with `csrfProtectionNeeded: true`)
    require the **double-submit cookie** pattern:

...

...
    rotated hourly.

...

    ---

...

...

...

    ---
...
servers:
...
components:
  securitySchemes:

    cookieAuth:
      type: apiKey
      in: cookie
      name: lumina_token
      description: |
        httpOnly JWT cookie set by `POST /expapi/v1/login` or the OAuth2 callback.
        Preferred auth method for browsers (immune to XSS). Sent automatically by
        the browser; non-browser clients should use `bearerAuth` instead.

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <jwt>` — legacy/alternative auth for non-browser
        clients (CLI, scripts, integration tests). Same JWT, same 1h expiry as
        the cookie.

    internalKey:
      type: apiKey
      in: header
      name: internal-key
      description: |
        Required on `/expapi/internal/*` routes. Constant-time compared against
        `INTERNAL_API_KEY`. Only the Discord bot should possess this key.

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on state-changing routes (`POST`/`PUT`/`DELETE`) that have
        `csrfProtectionNeeded: true`. The value must match the `csrf_token`
        cookie obtained from `GET /expapi/v1/csrf-token`.

    apiKeyQuery:
      type: apiKey
      in: query
      name: apiKey
...
paths:
...
  /expapi/oauth2/{provider}/auth/callback:
    get:
      tags: [OAuth2]
      summary: OAuth2 callback
      description: |
        Handles the OAuth2 provider redirect-back. Verifies the `state` JWT
        (must be ≤ 10 minutes old, with an allowed origin), exchanges the
        `code` for tokens, fetches the provider profile, and then dispatches
        based on the intent stored in `state`:

        - **link** — links the provider to `state.linkAccountId`, sets the
          `lumina_token` cookie, redirects to `{origin}/oauth/complete#...&linkedDiscord=true`.
        - **register** — creates a new account (or redirects with `email_exists`);
          for new accounts, attempts to seed `username`/`displayName` from the
          Discord profile (validated for syntax + availability).
        - **login** (default) — looks up by provider ID; falls back to verified
          email and auto-links; otherwise creates a new account automatically.

        Always redirects (no JSON) on success — the SPA reads the result from
        the redirect hash: `#isNewAccount=...&hasPassword=...`.

        For new Discord accounts, sets `discordOauth2Id`, `discordOauth2Token`,
        `discordOauth2RefreshToken`, `discordOauth2TokenExpiresAt`,
        `discordOauth2TokenScope`, etc.
      operationId: oauth2Callback
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [discord]
        - in: query
          name: code
          required: false
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
          description: Signed JWT produced by `/auth/start`.
        - in: query
          name: error
          required: false
          schema:
            type: string
          description: Provider-side error (e.g. `access_denied`).
      responses:
        '302':
          description: |
            Redirects to `{origin}/oauth/complete#isNewAccount=<bool>&hasPassword=<bool>`
            on success, or to `{origin}/login?oauthError=<code>` /
            `{origin}/register?oauthError=<code>` on failure.
          headers:
            Set-Cookie:
              description: |
                `lumina_token=<jwt>; HttpOnly; SameSite=Lax; Max-Age=3600` on success.
              schema:
                type: string
            Location:
              schema:
                type: string
        '400':
          description: Invalid or expired state.
          content:
            text/plain:
              schema:
                type: string
                example: Estado inválido ou expirado. Tente novamente.
        '404':
          description: Unknown provider (returned as plain text).
          content:
            text/plain:
              schema:
                type: string

  # ───────────────────────────────────────────────────────────────────────
  # USER
  # ───────────────────────────────────────────────────────────────────────
...